+++ src/components/core/layout/Visibility/Visibility-types.ts
@@ -1,0 +1,7 @@
+export interface ContainerProps {
+  /** - */
+  children: React.ReactNode;
+  desktop?: boolean;
+  tablet?: boolean;
+  mobile?: boolean;
+}